+++ js/technical/temp.js
@@ -153,7 +153,7 @@ function updateClickableTemp(layer)
 
 function setupBuyables(layer) {
 	for (id in layers[layer].buyables) {
-		if (!isNaN(id)) {
+		if (isPlainObject(layers[layer].buyables[id])) {
 			let b = layers[layer].buyables[id]
 			b.actualCostFunction = b.cost
 			b.cost = function(x) {