+++ changelog.md
@@ -1,5 +1,7 @@
 # The Modding Tree changelog:
 
+- You can now use non-numeric ids for upgrades, buyables, etc.
+
 # v2.5.5.2 - 5/12/21
 - Fixed a major issue with buyables.
 - Fixed a variety of tabFormat-related issues.